[Distributed][auto-parallel] Automatic Partitioning (#344)
This PR contains:

1. An ILP solver to find the optimal partitioning plan for the given
model, minimizing the communication cost while ensuring the parameters
not exceed the memory budget;
2. `hidet.distributed.partition`, which partitions the whole flow graph
according to the plan given by 1), and save the partitions to disk;
3. `hidet.distributed.load_partition`, which loads partition from the
disk to the desired device;
4. a launch script
5. a small resnet example(`example/distributed/resnet.py`). Which can be
run by `python -m hidet.distributed.launch [n_gpus] resnet.py`. The
memory budget has been set as 24MiB, which is less than the weight size
of a resnet18 model. It can help test tensor parallel

Several issues I plan to solve in the future
1. ILP solving is slow for large models like llama. We need to coalesce
nodes before sending the graph to ILP;
2. Multi processes on one machine running compilation in parallel often
trigger conflicts in local filesystem. We should implement a locking
mechanism in the future.